Add swiftlint and fix warning on code and generated code Rename Parser to Natalie and extract code from Storyboard to specific extension Edit macOS storyboard to add identity to view controllers
